# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blood Strike Scrim Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /violations:
    get:
      operationId: listViolations
      tags: [violations]
      summary: List all violations
      responses:
        "200":
          description: Violations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"
    post:
      operationId: createViolation
      tags: [violations]
      summary: Create a violation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateViolationBody"
      responses:
        "201":
          description: Created violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
  /violations/{id}:
    patch:
      operationId: updateViolation
      tags: [violations]
      summary: Update violation (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateViolationBody"
      responses:
        "200":
          description: Updated violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        bloodStrikeId:
          type: string
        role:
          type: string
          enum: [player, captain, admin]
        isBanned:
          type: boolean
        banExpiresAt:
          type: ["string", "null"]
        totalKills:
          type: integer
        totalWins:
          type: integer
        totalPoints:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - username
        - bloodStrikeId
        - role
        - isBanned
        - banExpiresAt
        - totalKills
        - totalWins
        - totalPoints
        - createdAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        bloodStrikeId:
          type: string
        role:
          type: string
          enum: [player, captain, admin]
      required:
        - username
        - password
        - bloodStrikeId
        - role
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        bloodStrikeId:
          type: string
        role:
          type: string
          enum: [player, captain, admin]
        isBanned:
          type: boolean
        banExpiresAt:
          type: ["string", "null"]
...
    Violation:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        scrimId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [no_show, rule_breaking, banned_weapon, other]
        description:
          type: string
        pointDeduction:
          type: integer
        status:
          type: string
          enum: [active, resolved]
        createdAt:
          type: string
      required:
        - id
        - teamId
        - teamName
        - userId
        - username
        - scrimId
        - type
        - description
        - pointDeduction
        - status
        - createdAt
    CreateViolationBody:
      type: object
      properties:
        teamId:
          type: integer
        userId:
          type: integer
        scrimId:
          type: integer
        type:
          type: string
          enum: [no_show, rule_breaking, banned_weapon, other]
        description:
          type: string
        pointDeduction:
          type: integer
      required:
        - teamId
        - type
        - description
        - pointDeduction
    UpdateViolationBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, resolved]
        pointDeduction:
          type: integer
...