# lib/api-zod/src/generated/types/updateViolationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Strike Scrim Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateViolationBodyStatus } from "./updateViolationBodyStatus";

export interface UpdateViolationBody {
  status?: UpdateViolationBodyStatus;
  pointDeduction?: number;
}