# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blood Strike Scrim Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  username: string;
  bloodStrikeId: string;
  role: UserRole;
  isBanned: boolean;
  /** @nullable */
  banExpiresAt: string | null;
  totalKills: number;
  totalWins: number;
  totalPoints: number;
  createdAt: string;
}